I noticed a problem with the autoconf style builds that shows up on 64 bit - some of the scripts had /usr/lib/mozldap6 in the PATH, and some had /usr/lib64/mozldap6.  All of them should have lib64.  What was happening is that make would create some script templates from the corresponding template .in file, and some it would just copy over (those that have both a template-script and a template-script.in).  The template-script files have hardcoded /usr/lib/ - it requires the conversion from template-script.in to template-script to substitute /usr/lib64/ instead.
To solve this problem, I touched all of the *.in files so that they would have a later date than their corresponding template file, so that make would create the template-script from template-script.in.  Make uses VPATH, and VPATH is used not only to look for dependencies, but also to find _targets_.  So if the .in file is older than its template file, make will find ../ldap/admin/src/scripts/template-foo (the target) is newer than ../ldap/admin/src/scripts/template-foo (the dependency) and will think it is up to date.  What should happen is that make should see that built/ldap/admin/src/scripts/template-foo does not exist, and create it from ../ldap/admin/src/scripts/template-foo.in, but apparently VPATH overrides that.

This problem will go away once we remove all of the template-script files and just create all of them from the template-script.in files.