Resolves: bug 160235 Bug Description: Add support for /etc/init scripts Reviewed by: nkinder (Thanks!) Fix Description: Add the new initscript.  The initscript is called $PACKAGE_NAME which by default is fedora-ds.  This script is created from wrappers/initscript.in, sed'd by the fixupcmd in Makefile.am during make install.  The way it works is this: service fedora-ds cmd will execute the cmd on all instances (found in /etc/fedora-ds by default).  service fedora-ds cmd instance will execute cmd on only that instance.  So if you have /etc/fedora-ds/slapd-foo /etc/fedora-ds/slapd-bar and you do service start fedora-ds it will start up both slapd-foo and slapd-bar.    If you do service start fedora-ds bar it will start up only slapd-bar.  If you do service start fedora-ds biff you will get an error message.    The initdir is platform specific (e.g. /etc/rc.d/init.d on linux, /etc/init.d on Solaris) so the definition was added to the platform dependent section of configure.ac.
The init script is explicitly branded, including the filename.    I needed to add
support to the autotool files so that we could change the name of the file.
Since package_name is defined when you use the AC_INIT macro in configure.ac,
we don't need to define it elsewhere (e.g. #define BRAND_DS).  So I added the
branding and other information to the autotool files, and changed
create_instance to use package_name instead of brand_ds to be consistent.
Having the package_name defined in much fewer places should make it much easier
to change in the future if necessary.

I also fixed a compiler warning in ldaprot.h.
Platforms tested: RHEL4, FC6
Flag Day: no
Doc impact: Yes.  We need to document how to use the initscript, and how to enable startup on boot - chkconfig fedora-ds on