Auto-generate SLAPI docs - first pass.
This starts the effort to be able to use Doxygen to
auto-generate the SLAPI documentation.

I started documenting everything in slapi-plugin.h
from the top down.  There is a TODO comment indicating
where the effort needs to be picked up from in the
header file.

To build the SLAPI docs, run "doxygen slapi.doxy" in
the top of the source tree.  The resulting doc files
will be in a subdirectory named "docs".  The build of
the docs still needs to be integrated with the rest of
the build.