507460 Access log could mistakenly report notes=U for VLV searches
Summary: Access log reports 'notes=U' for VLV indexed searches if there are no records to be found
Fix Description: VLV creates an empty IDL if no matched entries are found.  To do so, VLV code was calling idl_alloc with argument 0, which generated ALLID.  It's changed to call idl_alloc with 1.  It creates a normal empty IDL.