Add linked attributes plug-in.
This patch implements a linked attribute plug-in.  Details of the
plug-in features and design are available on the 389 wiki at:

  http://directory.fedoraproject.org/wiki/Linked_Attributes_Design

In addition, I encountered a memory leak when testing the new plug-in
with valgrind.  There was a fix to the dse add code for a double free
a few months back, which causes a leak in certain situations.  This
previous fix was for bz#489763.  The proper thing to do is to make
the dse backend add function consume the passed in entry upon success
and leave it for the caller to deal with upon failure.  This is the
way the back-ldbm add function works.