506137 ns-slapd hang while group aci performance testing
Bug description: If a group has more than 32767 members (max short),
a variable 'n' declared as short overflows.  The value is used to calculate an
array size to store group member info, which memory is not properly allocated
and it ends up crashing up the server.

Fix description: Replaced the problematic short variable type with integer.
Plus, the each member info was storing a pointer pointing to an element inside
of the array.  When the array is "realloc"ed, it's possible for the addresses
to be relocated.  To solve the problem, the new code stores the index of array
instead of the address.