Bug: 509401 - dnaNextValue not updated when dnaMaxValue set to -1
When "dnaMaxValue" is set to "-1" or omitted from a range configuration entry
(which defautls to "-1" internally), the "dnaNextValue" attribute is not
updated in the range configuration entry when a value is allocated from that
range.

We were only updating the configuration entry if the new nextvalue was >=
the maxval plus the interval (1).  We need to check if the maxval is -1
specifically, and update the config entry if so.