Add additional standard syntaxes.
This adds support for the following standard syntaxes, complete
with validation functions:

    Bit String
    Delivery Method
    Enhanced Guide
    Facsimile Telephone Number
    Fax
    Guide
    Name And Optional UID
    Printable String
    Teletex Terminal Identifier
    Telex Number

This patch does not change the schema to use any of these syntaxes
yet.  That will come when we update to the current versions of the
standard schema from the LDAP RFCs.

I also fixed an error in makefile.am where Setup.pm was listed
twice in perl_DATA.