Bug 514955 - Make DNA handle multiple mods
DNA doesn't handle multiple mods to a managed attribute
in the same modify operation properly.  If an operation
such as deleting a managed value triggers generation, we
aren't checking if another mod in the same operation is
actually adding a new value.  This triggers us to generate
a value when we really shouldn't.  The fix is to unset the
generate flag if we find a subsequent mod to the same
managed type.  It will be reset if we truly need to
generate a new value.