509472  db2index all does not reindex all the db backends correctly
db2index all (internally, called upgradedb) reads through the main db
id2entry.db# and reindex all the associated indexed attributes.  The
reindex borrows the import code where the entry id is newly assigned.
The new entry id's are connective.  On the other hand, entry id's of the
entries in the db to be reindexed are not.  The borrowed import code
assumes the entry id and the index of the fifo are tightly coupled and
the timing when the writing to and reading from the fifo are calculated
based upon the
assumption.

The assumption should have been revised so that the entry id which is
available up to is kept in ready_EID in the job structure and entry id from
each entry (entry->ep_id) is compared with ready_EID instead of ready_ID
that holds the sequential number.

Additionally, I eliminated unused variable "shift" from import_fifo_fetch.
Also, _dblayer_delete_instance_dir cleans up files and directories, recursively.