506786  Index maintenance mechanism causes wrong search results when modifying attributes with subtypes
When there are identical attribute value pairs except subtypes exist
in an entry, if one of the pairs are deleted, it should not affect the
index the attribute value is the key.

e.g.,
 mail: abc
 mail;en: abc
 mail;fr: xyz

 removing mail=abc or mail;en=abc, should not remove =abc from the
 mail.db#.

This fix uses the value array evals to determine if the equality key
in the index should be deleted or not.  The value array evals stores
the values of the attribute in the entry after the deletion is done.
If evals is empty, it means the to-be-deleted attribute value pair is
the only pair in the entry.  Thus, the equality key can be removed fom
the index.

If evals has values, then the to-be-deleted attribute (curr_attr,
which was retrieved from the old entry) value needs to be checked if
it's in evals or not.  If it is in evals, the equality key is still
used by other pair(s).  So, leave it.  Otherwise, the key can be
removed.

In the above example, let's assume removing mail=abc.  evals holds
{"abc", "xyz"}.  curr_attr abc is in evals, thus =abc will not be
removed.