506786 Index maintenance mechanism causes wrong search results when modifying attributes with subtypes
Andrey Ivanov (andrey.ivanov@polytechnique.fr) pointed out my previous
check-in for bug 506786 had an inefficient code.  To determine whether
to delete an equality index key or not, the code checks the key still
exists in the value array having the same attribute type.  The check
should be done as soon as one value is found in the value array instead
of checking through all of them.