Add centralized start/stop/restart scipts.
This adds centralized start, stop, and restart scripts
for ns-slapd.  These scripts live in the sbin directory
and will act upon all instances if an instance identifier
is not specified (similar to the init script).  The
instance specific scripts have been modified to call the
new centralized scripts.

The instance specific parameters needed by the new scripts
are located in the instance specific initconfig scripts,
which are now created by setup-ds.pl with values mapped
from the inf file.