Don't build policy module from Makefile.
This removes the Makefile rule that builds the SELinux policy
module.  The removed rule was only building and installing
the module for the targeted policy.  There are different base
policies (targeted, strict, mls) on different systems, so it
makes more sense to build the policy module from the spec file
where we can define the available base policy types for the
platform in question.

We still need a "--with-selinux" option to enable the SELinux
specific setup code as well as creating the policy .fc file
with the proper paths that are defined at build time.