Add ssf bind rule to access control plug-in.
This patch adds a new ssf bind rule keyword to the access control
plug-in.  This allows you to write ACIs that require a specific
level of encryption for the rule to apply.  The new keyword can
be used with '=', '!=', '<', '>', '<=' and '>=' comparators.

I added code that stores the SSF in effect for an operation into
the operation struct.  The value that we store is the higher of
the two between the SASL SSF and the SSL/TLS SSF.