Allow anonymous bind resource limits to be set.
This patch adds a new config setting named nsslapd-anonlimitsdn
that one can set to the DN of an entry containing the bind-based
resource limit attributes to use for operations performed by an
anonymous user.  This allows the defaults to still be used for
all other actual bound users who do not have any user specific
resource settings.

This implementation approach allows any resource limits that
are registered via the reslimit API to work with this anonymous
limits template entry.