Remove blank line from 00core.ldif.
There is an unnecessary blank line in 00core.ldif in the middle
of an entry.  This isn't causing any server issues since the line
does have a space, making it count at part of the previous attribute,
but it does through off Mozilla::LDAP::LDIF.  The blank line should
really be removed.