Bug 529258 - Make upgrade remove obsolete schema from 99user.ldif
Due to a previously fixed bug (bz 474254), standard schema may
have been duplicated in 99user.ldif.  When running an upgrade, we
need to check the 99user.ldif file to see if any obsolete standard
schema is defined there and remove it.

This patch makes upgrade backup the original 99user.ldif, and
scans it to see if any of the obsolete attributeTypes or
objectClasses attributes that we are removing from the standard
schema exist in 99user.ldif.  We trim the obsolete definitions
and write out a new 99user.ldif with the proper permissions.