529909 - Update SELinux policy for SASL GSSAPI
The dirsrv SELinux policy needs some changes to allow SASL GSSAPI
authentication to work.  We need to allow ns-slapd to read the
krb5.conf file and to create the in memory credentials cache.  The
kerberos libraries also attempt to open the krb5.conf in write mode,
so we need to prevent those attempts from being audited.