387681 - Fix errors in mapping AD tombstones
The AD tombstone mapping code is not behaving correctly if a
cn contains a comma (such as a "last, first" type value).  The
code is supposed to locate the first ":" in the tombstone DN,
then scan for the first "," after that.  Everything between is
the GUID.  The problem is that the code is starting at the
beginning of the string when searching for the "," instead of
starting at the ":" that was previously found.  This causes the
"," in the cn to be found instead, which makes us fail to find
the GUID.

The fix is to simply start searching for the "," from the ":"
in the tombstone DN.