Allow plugin execution order to be defined
This patch adds support for assigning a precedence to each plugin.
This precedence value is used to determine what order the plugins
should be called in within a given plugin type.

The precedence value is an integer between 1 and 99.  The value is
set in the nsslapd-pluginPrecedence attribute in a plugin config
entry in "cn=config".  If a plugin does not have a precedence
value set, the plugin is treated as if it has a precedence of 50.

For more details, see the feature design document at:

    http://directory.fedoraproject.org/wiki/Plugin_Ordering

I also removed an incorrect log message in the linked attrs plugin
that I noticed while debugging this patch.